% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astropic-package.R
\docType{package}
\name{astropic-package}
\alias{astropic}
\alias{astropic-package}
\title{astropic: Connection to APOD API}
\description{
This package aims to make calling the APOD API easy and fun so you can get straight to the cool science!
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/eringrand/astropic/issues}
}

}
\author{
\strong{Maintainer}: Erin Grand \email{eringrand@gmail.com}

}
\keyword{internal}
