% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apod.R
\name{get_apod}
\alias{get_apod}
\title{get_apod}
\usage{
get_apod(start_date = NULL, end_date = NULL)
}
\arguments{
\item{start_date}{starting date you'd like to look at}

\item{end_date}{end date you'd like to look at, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
