% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hist_apod}
\alias{hist_apod}
\title{Historical APOD image data from 2007 - 2025}
\format{
A data frame with 7304 rows and 8 variables:
\describe{
\item{copyright}{image copyright, if not public domain}
\item{date}{Date of the image}
\item{service_version}{unknown}
\item{explanation}{The supplied text explanation of the image, may have some HTML text still in it}
\item{hdurl}{The url for the high-resolution image}
\item{media_type}{The type of media returned. May either be 'image' or 'video' depending on content}
\item{title}{The title of image}
\item{url}{The url of image}
}
}
\usage{
hist_apod
}
\description{
Historical APOD image data from 2007 - 2025
}
\keyword{datasets}
