% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apod.R
\name{get_apod}
\alias{get_apod}
\title{get_apod}
\usage{
get_apod(query = NULL, print = FALSE)
}
\arguments{
\item{query}{Query you want
The query parameters are described on the \href{https://github.com/nasa/apod-api}{APOD API Github page}
\itemize{
\item \code{date} A string in YYYY-MM-DD format indicating the date of the APOD image (example: 2014-11-03). Defaults to today's date. Must be after 1995-06-16, the first day an APOD picture was posted. There are no images for tomorrow available through this API.
\item \code{concept_tags} A Boolean True|False indicating whether concept tags should be returned with the rest of the response. The concept tags are not necessarily included in the explanation, but rather derived from common search tags that are associated with the description text. (Better than just pure text search.) Defaults to False.
\item \code{count} A positive integer, no greater than 100. If this is specified then count randomly chosen images will be returned in a JSON array. Cannot be used in conjunction with date or \code{start_date} and \code{end_date}.
\item \code{start_date} A string in YYYY-MM-DD format indicating the start of a date range. All images in the range from \code{start_date} to end_date will be returned in a JSON array. Cannot be used with date.
\item \code{end_date} A string in YYYY-MM-DD format indicating that end of a date range. If \code{start_date} is specified without an \code{end_date} then \code{end_date} defaults to the current date.
\item \code{thumbs} A Boolean parameter True|False indicating whether the API should return a thumbnail image URL for video files. If set to True, the API returns URL of video thumbnail. If an APOD is not a video, this parameter is ignored.
}}

\item{print}{if print = TRUE: Print out how where you are with the APOD API rate limit. Defaults to not printing anything.}
}
\value{
Returns a \code{tibble} containing
\itemize{
\item \code{date} Date of image. Included in response because of default values.
\item \code{explanation} The supplied text explanation of the image.
\item \code{hdurl} The URL for any high-resolution image for that day. Will be omitted in the response IF it does not exist originally at APOD.
\item \code{media_type} The type of media (data) returned. May either be 'image' or 'video' depending on content.
\item \code{service_version} The service version used.
\item \code{title} The title of the image.
\item \code{url} The URL of the APOD image or video of the day.
Others:
\item \code{resource} A dictionary describing the image_set or planet that the response illustrates, completely determined by the structured endpoint.
\item \code{concept_tags} A Boolean reflection of the supplied option. Included in response because of default values.
\item \code{concepts} The most relevant concepts within the text explanation. Only supplied if concept_tags is set to True.
\item \code{thumbnail_url} The URL of thumbnail of the video.
\item \code{copyright} The name of the copyright holder.
If there is an error with any images, there will also be an error column.
}
}
\description{
This function retrieves data from the APOD API.
}
\examples{
get_apod()
}
