% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apod.R
\name{get_apod}
\alias{get_apod}
\title{get_apod}
\usage{
get_apod(query = NULL, print = FALSE)
}
\arguments{
\item{query}{Query you want}

\item{print}{if print = TRUE: Print out how where you are with the APOD API rate limit
Defaults to not printing anything.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
